added all requestModels

+++ Findexium.Domain/Models/CurvePoint.cs
@@ -7,7 +7,7 @@ namespace Findexium.Domain.Models
         // TODO: Map columns in data table CURVEPOINT with corresponding fields
 
         public int Id { get; set; }
-        public byte? CurveId { get; set; }
+        public int? CurveId { get; set; }
         public DateTime? AsOfDate { get; set; }
         public double? Term { get; set; }
         public double? CurvePointValue { get; set; }
